backEnd/PythonAPI/requirements.txt
Django==2.2.5
djangorestframework==3.10.3
keras==2.2.4
opencv-python==4.1.1.26
Pillow==5.3.0
setuptools==41.0.0
tensorflow==1.14.0
